Update retrofit, support libraries, gradle plugin
