+++ app/src/main/java/com/genymobile/pr/net/IssueCallback.java
@@ -5,17 +5,17 @@ import com.genymobile.pr.bus.IssueRetrievedEvent;
 import com.genymobile.pr.model.Issue;
 
 import android.util.Log;
-import retrofit.Callback;
-import retrofit.Response;
-import retrofit.Retrofit;
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
 public class IssueCallback implements Callback<Issue> {
 
     private static final String TAG = IssueCallback.class.getSimpleName();
 
     @Override
-    public void onResponse(Response<Issue> response, Retrofit retrofit) {
-        if (response.isSuccess()) {
+    public void onResponse(Call<Issue> call, Response<Issue> response) {
+        if (response.isSuccessful()) {
             BusProvider.getInstance().post(new IssueRetrievedEvent(response.body()));
         } else {
             Log.e(TAG, "Couldn't load issue : " + response.message());
@@ -24,7 +24,7 @@ public class IssueCallback implements Callback<Issue> {
     }
 
     @Override
-    public void onFailure(Throwable t) {
+    public void onFailure(Call<Issue> call, Throwable t) {
         Log.e(TAG, "Couldn't load issue", t);
         //TODO handle error
     }
