+++ app/src/main/java/com/genymobile/pr/net/PullRequestsCallback.java
@@ -1,16 +1,15 @@
 package com.genymobile.pr.net;
 
-import android.util.Log;
-
+import com.genymobile.pr.bus.BusProvider;
 import com.genymobile.pr.bus.LoadingErrorEvent;
 import com.genymobile.pr.bus.NetworkErrorEvent;
-import com.genymobile.pr.model.PullRequest;
-import com.genymobile.pr.bus.BusProvider;
 import com.genymobile.pr.bus.PullRequestsRetrievedEvent;
+import com.genymobile.pr.model.PullRequest;
 
-import retrofit.Callback;
-import retrofit.Response;
-import retrofit.Retrofit;
+import android.util.Log;
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
 import java.util.List;
 
@@ -18,8 +17,8 @@ public class PullRequestsCallback implements Callback<List<PullRequest>> {
     private static final String TAG = PullRequestsCallback.class.getSimpleName();
 
     @Override
-    public void onResponse(Response<List<PullRequest>> response, Retrofit retrofit) {
-        if (response.isSuccess()) {
+    public void onResponse(Call<List<PullRequest>> call, Response<List<PullRequest>> response) {
+        if (response.isSuccessful()) {
             BusProvider.getInstance().post(new PullRequestsRetrievedEvent(response.body()));
         } else {
             Log.e(TAG, "Couldn't load pull requests : " + response.message());
@@ -28,7 +27,7 @@ public class PullRequestsCallback implements Callback<List<PullRequest>> {
     }
 
     @Override
-    public void onFailure(Throwable t) {
+    public void onFailure(Call<List<PullRequest>> call, Throwable t) {
         Log.e(TAG, "Couldn't load pull requests", t);
         BusProvider.getInstance().post(new NetworkErrorEvent());
     }
