+++ app/src/main/java/com/genymobile/pr/net/ReposCallback.java
@@ -1,25 +1,24 @@
 package com.genymobile.pr.net;
 
-import android.util.Log;
-
 import com.genymobile.pr.bus.BusProvider;
 import com.genymobile.pr.bus.LoadingErrorEvent;
 import com.genymobile.pr.bus.NetworkErrorEvent;
 import com.genymobile.pr.bus.ReposRetrievedEvent;
 import com.genymobile.pr.model.Repo;
 
-import java.util.List;
+import android.util.Log;
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
-import retrofit.Callback;
-import retrofit.Response;
-import retrofit.Retrofit;
+import java.util.List;
 
 public class ReposCallback implements Callback<List<Repo>> {
     private static final String TAG = ReposCallback.class.getSimpleName();
 
     @Override
-    public void onResponse(Response<List<Repo>> response, Retrofit retrofit) {
-        if (response.isSuccess()) {
+    public void onResponse(Call<List<Repo>> call, Response<List<Repo>> response) {
+        if (response.isSuccessful()) {
             BusProvider.getInstance().post(new ReposRetrievedEvent(response.body()));
         } else {
             Log.e(TAG, "Couldn't load repos : " + response.message());
@@ -28,7 +27,7 @@ public class ReposCallback implements Callback<List<Repo>> {
     }
 
     @Override
-    public void onFailure(Throwable t) {
+    public void onFailure(Call<List<Repo>> call, Throwable t) {
         Log.e(TAG, "Couldn't load repos", t);
         BusProvider.getInstance().post(new NetworkErrorEvent());
     }
